fix(desktop): folder-move descendant events + chunk_concurrency i64 (#151, #169)
#151: when a directory is moved/renamed, the OS emits Remove+Create
events for every descendant in addition to the directory's own events.
The event blocker only covered the directory's exact paths, so child
removes were propagated as remote deletes — children landed in the
recycle bin while the folder moved without them.

- EventBlocker: TTL-scoped path-prefix suppression (register_prefix);
  rename() suppresses Remove(source/*) + Create(target/*) for 15s around
  the remote call, covering the in-flight window
- process_fs_events: deterministic group order (name->create->modify->
  remove) instead of HashMap order so same-batch removes observe the
  post-move world
- tests: prefix descendant/self/sibling matching, expiry, exact-consume

#169: chunk_concurrency parsed as i32 overflowed on values above
i32::MAX, failing serde for the whole listing response and rendering
the sync root unusable. Widened to i64; consumer already clamps max(1).

+++ desktop/crates/cloudreve-api/src/models/explorer.rs
@@ -193,7 +193,7 @@ pub struct StoragePolicy {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub children: Option<Vec<StoragePolicy>>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub chunk_concurrency: Option<i32>,
+    pub chunk_concurrency: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub encryption: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
